add initial description in asciidoc
= Internet partagé configuration and scripts

== Purpose

This repository contains all essential information about the network topology, equipment, configuration and tools needed to set up, run and maintain the in-house network of the association Éco du boiron.

We share all non-sensible information about our technical solution here for our future administators as well as a reference for the current work in progress. We hope other people can learn from it or notify us of faults or security holes in our concept.